+++ src/main/java/kitchenpos/dao/OrderTableDao.java
@@ -2,7 +2,7 @@ package kitchenpos.dao;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
 public interface OrderTableDao {
     OrderTable save(OrderTable entity);
