+++ src/main/java/kitchenpos/dao/TableGroupDao.java
@@ -2,7 +2,7 @@ package kitchenpos.dao;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.order.domain.TableGroup;
 
 public interface TableGroupDao {
     TableGroup save(TableGroup entity);
