+++ src/test/java/kitchenpos/acceptance/TableGroupAcceptanceTest.java
@@ -6,8 +6,8 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.AcceptanceTest;
 import kitchenpos.acceptance.helper.KitchenPosBehaviors;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.TableGroup;
 import kitchenpos.fixture.OrderTableFixtureFactory;
 import kitchenpos.fixture.TableGroupFixtureFactory;
 import org.assertj.core.util.Lists;
