+++ src/test/java/kitchenpos/application/helper/ServiceTestHelper.java
@@ -6,13 +6,13 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.IntStream;
 import kitchenpos.application.OrderService;
-import kitchenpos.application.TableGroupService;
-import kitchenpos.application.TableService;
+import kitchenpos.order.application.TableGroupService;
+import kitchenpos.order.application.TableService;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.TableGroup;
 import kitchenpos.fixture.MenuFixtureFactory;
 import kitchenpos.fixture.MenuGroupFixtureFactory;
 import kitchenpos.fixture.OrderFixtureFactory;
