+++ src/test/java/kitchenpos/fixture/OrderTableFixtureFactory.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
 public class OrderTableFixtureFactory {
     private OrderTableFixtureFactory() {
